backend/requirements-base.txt
# Core Django dependencies
Django==4.2.7
channels==4.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
daphne==4.0.0

# Database
# SQLite is included with Python

# Data processing
numpy==1.24.4
scikit-learn==1.3.2
pandas==2.0.3

# Utilities
tabulate==0.9.0
tqdm==4.66.1
requests==2.31.0

# Async support
asyncio==3.4.3
websockets==12.0